Time-Piece: Use foldEQ_locale() if available
This core function (available since 5.13.2) is thread-safe and knows
about Perl internals, so is preferable to the similar libc function,
which is now used only as a fallback.  This commit also bomb proofs the
code by adding an additional fallback, specified in C89, which isn't a
great substitute, but far better than nothing.